backend/requirements.txt
# Flask para el backend
Flask==2.3.2

# Flask-CORS para manejar CORS
Flask-Cors==3.0.10

# Flask-JWT-Extended para autenticación con JWT
Flask-JWT-Extended==4.4.4

# Flask-SQLAlchemy para trabajar con bases de datos de manera ORM
Flask-SQLAlchemy==2.5.1

# PyMySQL para conexión a MySQL
PyMySQL==1.0.3

# python-dotenv para manejar variables de entorno
python-dotenv==1.0.0

# Werkzeug para manejar utilidades WSGI y de seguridad
Werkzeug==2.3.6

# Conector alternativo para MySQL
mysql-connector-python==8.0.33

# Para exportar PDF
xhtml2pdf==0.2.11
